Moved to stdplus: vector_mutexed_obj with_mutex
